requirements.txt
certifi==2025.7.9
charset-normalizer==3.4.2
idna==3.10
numpy==2.3.1
pandas==2.3.1
pycryptodome==3.23.0
pyqt6==6.9.1
pyqt6-qt6==6.9.1
pyqt6-sip==13.10.2
pyside6==6.9.1
pyside6-addons==6.9.1
pyside6-essentials==6.9.1
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
requests==2.32.4
shiboken6==6.9.1
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1

fastapi~=0.116.1
boto3~=1.40.21
uvicorn~=0.35.0
python-dotenv~=1.1.1